Moving activities back to the schedule package
