+++ test/org/sagebionetworks/bridge/models/schedules/TaskSchedulerTest.java
@@ -17,7 +17,6 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.sagebionetworks.bridge.TestConstants;
-import org.sagebionetworks.bridge.models.tasks.Activity;
 
 import com.newrelic.agent.deps.com.google.common.collect.Maps;
