+++ test/org/sagebionetworks/bridge/validators/ScheduleValidatorTest.java
@@ -10,9 +10,9 @@ import org.junit.Test;
 import org.sagebionetworks.bridge.TestConstants;
 import org.sagebionetworks.bridge.exceptions.InvalidEntityException;
 import org.sagebionetworks.bridge.json.DateUtils;
+import org.sagebionetworks.bridge.models.schedules.Activity;
 import org.sagebionetworks.bridge.models.schedules.Schedule;
 import org.sagebionetworks.bridge.models.schedules.ScheduleType;
-import org.sagebionetworks.bridge.models.tasks.Activity;
 import org.sagebionetworks.bridge.models.tasks.ActivityType;
 
 public class ScheduleValidatorTest {
